fix:No.23_商品登録時に金額がDouble型で登録されない不具合を修正

+++ src/main/java/com/example/model/Product.java
@@ -48,7 +48,7 @@ public class Product extends TimeEntity implements Serializable {
 	private Integer height;
 
 	@Column(name = "price", nullable = false)
-	private Integer price;
+	private Double price;
 
 	@Column(name = "tax_type", nullable = false)
 	private Integer taxType;
@@ -64,7 +64,7 @@ public class Product extends TimeEntity implements Serializable {
 		this.setCode(form.getCode());
 		this.setWeight(form.getWeight());
 		this.setHeight(form.getHeight());
-		this.setPrice(form.getPrice());
+		this.setPrice((double)form.getPrice());
 		this.setTaxType(form.getTaxType());
 	}
